+++ middleware.ts
@@ -1,7 +1,7 @@
 import { withAuth } from "next-auth/middleware";
 import { NextResponse } from "next/server";
 
-const openRoutes = ["/", "/api/stripe/webhook"];
+const openRoutes = ["/", "/api/stripe/webhook", "/api/metrics"];
 
 export default withAuth(
   function middleware(req) {